requirements.txt
faiss-cpu
PyPDF2
Flask==3.0.3
python-dotenv==1.0.1
gunicorn
httpx

# Phase 3 — embeddings
sentence-transformers==2.7.0
numpy==1.26.4
requests

pdfminer.six

beautifulsoup4
